Set Human Readable Representation for Tournament Model
Set up a __str__ method for the tournament model to make a Readable
representation of the tournament instance,
As the tournament currently only contains a start date and an end date,
the representation of the tournament is the duration of the tournament.
(e.g: if start date is set to 2024-04-28 and end_date is set to
2024-04-30, the __str__ method should return a string
"2024-04-28 - 2024-04-30")
A simple test case to assert this is included.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>