Add Players to Teams
Players are currently defined simply as users of the app as well.
We have a many-to-many relationship between Teams and Players through
a TeamMember model. This allows people to potentially be members of
multiple teams. Team-specific data of the player such as jersey number
etc. can be stored on the intermediate TeamMember model.

Currently, all data is managed via the django admin panel. This does not
allow for proper management such as add a team including members
directly from the same form, creating new users where necessary.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>