Make Teams Participate in Tournaments
Teams have a many-to-many relationship with tournaments through an
intermediate Participant model.
This allows for teams to run in several tournaments and each tournament
allowing the participation of many teams.
Additionally, we can store information about teams in a given tournament
in this intermediate Model when required.

The admin interface is adjusted so that teams can be added to
tournaments inline within the tournament edit form.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>