Add Django-Extensions and Pydotplus to the Poetry Project.
Django-Extensions is a third-party application extending django
functionality. A variety of manage.py functions are added.
Pydotplus is added to make use of the django extensions graph_models
function.
This function allows to generate a dotfile representation of the
database schema (UML diagram) and output this graphic in different
formats.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>