Add Games Model to Tournaments
A tournament consists of many games.
Each game is played by two teams which are participants of the
tournament. The teams are classified as home team and away team.
For each game we currently record only the score as home goals and
away goals. The score is needed to later calculate points for the
leaderboard. Leaderboard stats will used to generate game plans.
Depending on tournament mode, games might have to be sorted into rounds.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>