Display Leaderboard Statistics in Tournament Detail View
Add a function to calculate the number of games played in the tournament
for each participant on the Participant model.

Add a function to calculate the points scored in the tournament for each
participant. Points calculation is based on the 3-1-0 system:
Winning a game scores 3 points, drawing scores 1 point and losing scores
no points.

The calculated number of games and points scored are displayed in the
Tournament detail view in a table.

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>