Define Tournament Rounds
A Tournament is supposed to be played in rounds.
We implement a Rounds table with the following information:
Which tournament does the round belong to?
Which round of the tournament is it (first, second...)?
What Round Type is it (Standard or Final for now)?
When is the round supposed to start?
How long is each half of a game in the round supposed to be?
How long is the half time break supposed to be?
How long is the break between games supposed to be?

We have a one-to many relationship to tournaments (Many rounds per
tournament, but each round only belongs to one tournament).
We also have a one-to many relationship to games (each game belongs to
one round, but each round can have many games).

Signed-off-by: Daniel Zaumsegel <daniel.zaumsegel@gmail.com>